Divide ProductPage into components
